+++ server/src/main/java/com/ca/lsp/core/cobol/engine/CobolLanguageEngine.java
@@ -13,14 +13,14 @@
  */
 package com.ca.lsp.core.cobol.engine;
 
-import com.broadcom.lsp.domain.common.model.Position;
 import com.ca.lsp.core.cobol.model.ExtendedDocument;
+import com.ca.lsp.core.cobol.model.Locality;
 import com.ca.lsp.core.cobol.model.ResultWithErrors;
 import com.ca.lsp.core.cobol.model.SyntaxError;
 import com.ca.lsp.core.cobol.parser.CobolLexer;
 import com.ca.lsp.core.cobol.parser.CobolParser;
 import com.ca.lsp.core.cobol.preprocessor.CobolPreprocessor;
-import com.ca.lsp.core.cobol.preprocessor.sub.util.impl.PositionMappingUtils;
+import com.ca.lsp.core.cobol.preprocessor.sub.util.impl.LocalityMappingUtils;
 import com.ca.lsp.core.cobol.semantics.SemanticContext;
 import com.ca.lsp.core.cobol.strategy.CobolErrorStrategy;
 import com.ca.lsp.core.cobol.visitor.CobolVisitor;
@@ -37,6 +37,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
+import java.util.function.Predicate;
+import java.util.stream.Stream;
 
 import static java.util.stream.Collectors.toList;
 
@@ -87,8 +89,8 @@ public class CobolLanguageEngine {
 
     CobolParser.StartRuleContext tree = parser.startRule();
 
-    Map<Token, Position> positionMapping =
-        PositionMappingUtils.createPositionMapping(
+    Map<Token, Locality> positionMapping =
+        LocalityMappingUtils.createPositionMapping(
             tokens.getTokens(), extendedDocument.getDocumentMapping(), documentUri);
 
     CobolVisitor visitor =
@@ -97,21 +99,47 @@ public class CobolLanguageEngine {
 
     accumulatedErrors.addAll(finalizeErrors(listener.getErrors(), positionMapping));
     accumulatedErrors.addAll(visitor.getErrors());
+    accumulatedErrors.addAll(
+        collectErrorsForCopybooks(accumulatedErrors, extendedDocument.getCopyStatements()));
 
     return new ResultWithErrors<>(visitor.getSemanticContext(), accumulatedErrors);
   }
 
   @Nonnull
   private List<SyntaxError> finalizeErrors(
-      @Nonnull List<SyntaxError> errors, @Nonnull Map<Token, Position> mapping) {
+      @Nonnull List<SyntaxError> errors, @Nonnull Map<Token, Locality> mapping) {
     return errors.stream()
         .map(convertError(mapping))
-        .filter(it -> it.getPosition() != null)
+        .filter(it -> it.getLocality() != null)
         .collect(toList());
   }
 
   @Nonnull
-  private Function<SyntaxError, SyntaxError> convertError(@Nonnull Map<Token, Position> mapping) {
-    return err -> err.toBuilder().position(mapping.get(err.getOffendedToken())).build();
+  private Function<SyntaxError, SyntaxError> convertError(@Nonnull Map<Token, Locality> mapping) {
+    return err -> err.toBuilder().locality(mapping.get(err.getOffendedToken())).build();
+  }
+
+  private List<SyntaxError> collectErrorsForCopybooks(
+      List<SyntaxError> syntaxErrors, Map<String, Locality> copyStatements) {
+    return syntaxErrors.stream()
+        .filter(shouldRaise())
+        .map(err -> raiseError(err, copyStatements))
+        .flatMap(List::stream)
+        .collect(toList());
+  }
+
+  private List<SyntaxError> raiseError(
+      SyntaxError syntaxError, Map<String, Locality> copyStatements) {
+    return Stream.of(syntaxError)
+        .filter(shouldRaise())
+        .map(err -> err.toBuilder().locality(copyStatements.get(err.getLocality().getCopybookId())))
+        .map(SyntaxError.SyntaxErrorBuilder::build)
+        .map(err -> Stream.concat(raiseError(err, copyStatements).stream(), Stream.of(err)))
+        .flatMap(Function.identity())
+        .collect(toList());
+  }
+
+  private Predicate<SyntaxError> shouldRaise() {
+    return err -> err.getLocality().getCopybookId() != null;
   }
 }