+++ server/src/main/java/com/ca/lsp/core/cobol/visitor/CobolVisitor.java
@@ -14,13 +14,13 @@
 
 package com.ca.lsp.core.cobol.visitor;
 
-import com.broadcom.lsp.domain.common.model.Position;
+import com.ca.lsp.core.cobol.model.Locality;
 import com.ca.lsp.core.cobol.model.SyntaxError;
 import com.ca.lsp.core.cobol.model.Variable;
 import com.ca.lsp.core.cobol.parser.CobolParser.*;
 import com.ca.lsp.core.cobol.parser.CobolParserBaseVisitor;
-import com.ca.lsp.core.cobol.semantics.CobolVariableContext;
 import com.ca.lsp.core.cobol.preprocessor.sub.util.impl.PreprocessorStringUtils;
+import com.ca.lsp.core.cobol.semantics.CobolVariableContext;
 import com.ca.lsp.core.cobol.semantics.NamedSubContext;
 import com.ca.lsp.core.cobol.semantics.SemanticContext;
 import com.ca.lsp.core.cobol.semantics.SubContext;
@@ -32,17 +32,18 @@ import org.antlr.v4.runtime.CommonTokenStream;
 import org.antlr.v4.runtime.RuleContext;
 import org.antlr.v4.runtime.Token;
 import org.eclipse.lsp4j.DocumentSymbol;
+import org.eclipse.lsp4j.Location;
+import org.eclipse.lsp4j.Range;
 
 import javax.annotation.Nonnull;
-import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
 import java.util.Map;
+import java.util.Optional;
 
-import static com.ca.lsp.core.cobol.semantics.outline.OutlineNodeNames.*;
 import static com.ca.lsp.core.cobol.model.ErrorSeverity.INFO;
 import static com.ca.lsp.core.cobol.model.ErrorSeverity.WARNING;
+import static com.ca.lsp.core.cobol.semantics.outline.OutlineNodeNames.*;
 import static java.util.Optional.ofNullable;
 import static java.util.stream.Collectors.joining;
 
@@ -62,6 +63,7 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
       "The following token cannot be on the same line as a DECLARATIVE token: ";
   private static final String INVALID_DEF_MSG = "Invalid definition for: ";
   private static final String MISSPELLED_WORD = "A misspelled word, maybe you want to put ";
+  private static final String PROGRAM_ID_ISSUE_MSG = "There is an issue with PROGRAM-ID paragraph";
 
   @Getter private List<SyntaxError> errors = new ArrayList<>();
 
@@ -73,13 +75,13 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
   private NamedSubContext copybooks;
   private CommonTokenStream tokenStream;
   private OutlineTreeBuilder outlineTreeBuilder;
-  private Map<Token, Position> positionMapping;
+  private Map<Token, Locality> positionMapping;
 
   public CobolVisitor(
       @Nonnull String documentUri,
       @Nonnull NamedSubContext copybooks,
       @Nonnull CommonTokenStream tokenStream,
-      @Nonnull Map<Token, Position> positionMapping) {
+      @Nonnull Map<Token, Locality> positionMapping) {
     this.copybooks = copybooks;
     this.positionMapping = positionMapping;
     this.tokenStream = tokenStream;
@@ -113,11 +115,15 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
 
   @Override
   public Class visitProgramIdParagraph(ProgramIdParagraphContext ctx) {
-    if (ctx.programName() != null) {
-      programName = ctx.programName().getText();
-      outlineTreeBuilder.renameProgram(programName, ctx);
-      outlineTreeBuilder.addNode(PROGRAM_ID_PREFIX + programName, NodeType.PROGRAM_ID, ctx);
-    }
+    ofNullable(ctx.programName())
+        .map(RuleContext::getText)
+        .map(PreprocessorStringUtils::trimQuotes)
+        .ifPresent(
+            name -> {
+              programName = name;
+              outlineTreeBuilder.renameProgram(name, ctx);
+              outlineTreeBuilder.addNode(PROGRAM_ID_PREFIX + name, NodeType.PROGRAM_ID, ctx);
+            });
     return visitChildren(ctx);
   }
 
@@ -209,7 +215,10 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
     Token endToken = ctx.END().getSymbol();
 
     if (firstDeclLine == declarativeBody.getLine()) {
-      throwException(declarativeBody.getText(), getPosition(declarativeBody), DECLARATIVE_SAME_MSG);
+      getLocality(declarativeBody)
+          .ifPresent(
+              locality ->
+                  throwException(declarativeBody.getText(), locality, DECLARATIVE_SAME_MSG));
     }
 
     areaAWarning(firstDeclarative);
@@ -319,8 +328,9 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
 
   @Override
   public Class visitParagraphName(ParagraphNameContext ctx) {
-    Position position = getPosition(ctx.getStart());
-    if (position != null) paragraphs.define(ctx.getText().toUpperCase(), position);
+    getLocality(ctx.getStart())
+        .map(Locality::toLocation)
+        .ifPresent(it -> paragraphs.define(ctx.getText().toUpperCase(), it));
     return visitChildren(ctx);
   }
 
@@ -335,8 +345,11 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
     ofNullable(ctx.dataName1())
         .ifPresent(
             variable ->
-                defineVariable(levelNumber, variable.getText(), getPosition(variable.getStart())));
-    String name = Optional.ofNullable(ctx.dataName1()).map(RuleContext::getText).orElse(FILLER_NAME);
+                getLocality(variable.getStart())
+                    .map(Locality::toLocation)
+                    .ifPresent(
+                        location -> defineVariable(levelNumber, variable.getText(), location)));
+    String name = ofNullable(ctx.dataName1()).map(RuleContext::getText).orElse(FILLER_NAME);
     int level = Integer.parseInt(levelNumber);
     outlineTreeBuilder.addVariable(level, name, getDataDescriptionNodeType(ctx), ctx);
     return visitChildren(ctx);
@@ -361,8 +374,12 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
     ofNullable(ctx.dataName1())
         .ifPresent(
             variable ->
-                defineVariable(levelNumber, variable.getText(), getPosition(variable.getStart())));
-    outlineTreeBuilder.addVariable(CobolVariableContext.LEVEL_66, ctx.dataName1().getText(), NodeType.FIELD_66, ctx);
+                getLocality(variable.getStart())
+                    .map(Locality::toLocation)
+                    .ifPresent(
+                        location -> defineVariable(levelNumber, variable.getText(), location)));
+    outlineTreeBuilder.addVariable(
+        CobolVariableContext.LEVEL_66, ctx.dataName1().getText(), NodeType.FIELD_66, ctx);
     return visitChildren(ctx);
   }
 
@@ -372,20 +389,29 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
     ofNullable(ctx.dataName1())
         .ifPresent(
             variable ->
-                defineVariable(levelNumber, variable.getText(), getPosition(variable.getStart())));
-    outlineTreeBuilder.addVariable(CobolVariableContext.LEVEL_88, ctx.dataName1().getText(), NodeType.FIELD_88, ctx);
+                getLocality(variable.getStart())
+                    .map(Locality::toLocation)
+                    .ifPresent(
+                        location -> defineVariable(levelNumber, variable.getText(), location)));
+    outlineTreeBuilder.addVariable(
+        CobolVariableContext.LEVEL_88, ctx.dataName1().getText(), NodeType.FIELD_88, ctx);
     return visitChildren(ctx);
   }
 
-  private void defineVariable(String level, String name, Position position) {
-    if (position == null) return;
-    variables.define(new Variable(level, name), position);
+  private void defineVariable(String level, String name, @Nonnull Location location) {
+    variables.define(new Variable(level, name), location);
+  }
+
+  private void addUsage(SubContext<?> langContext, String name, @Nonnull Location location) {
+    langContext.addUsage(name.toUpperCase(), location);
   }
 
   @Override
   public Class visitParagraphNameUsage(ParagraphNameUsageContext ctx) {
     String name = ctx.getText().toUpperCase();
-    addUsage(paragraphs, name, getPosition(ctx.getStart()));
+    getLocality(ctx.getStart())
+        .map(Locality::toLocation)
+        .ifPresent(it -> addUsage(paragraphs, name, it));
     return visitChildren(ctx);
   }
 
@@ -402,9 +428,9 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
   }
 
   private void checkForVariable(String variable, QualifiedDataNameFormat1Context ctx) {
-    Position position = getPosition(ctx.getStart());
-    checkVariableDefinition(variable, position);
-    addUsage(variables, variable, position);
+    Optional<Locality> locality = getLocality(ctx.getStart());
+    locality.ifPresent(it -> checkVariableDefinition(variable, it));
+    locality.map(Locality::toLocation).ifPresent(it -> addUsage(variables, variable, it));
 
     if (ctx.qualifiedInData() != null) {
       iterateOverQualifiedDataNames(ctx, variable);
@@ -421,23 +447,23 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
 
       String parent = context.getText().toUpperCase();
       Token parentToken = context.getStart();
-      Position parentPosition = getPosition(parentToken);
+      Optional<Locality> parentLocality = getLocality(parentToken);
 
-      checkVariableDefinition(parent, parentPosition);
+      parentLocality.ifPresent(it -> checkVariableDefinition(parent, it));
       checkVariableStructure(parent, child, childToken, parentToken);
 
       childToken = parentToken;
       child = parent;
 
-      addUsage(variables, child, parentPosition);
+      Location parentLocation = parentLocality.map(Locality::toLocation).orElse(null);
+      if (parentLocation != null) addUsage(variables, child, parentLocation);
     }
   }
 
-  private void throwException(String wrongToken, @Nullable Position position, String message) {
-    if (position == null) return;
+  private void throwException(String wrongToken, @Nonnull Locality locality, String message) {
     SyntaxError syntaxError =
         SyntaxError.syntaxError()
-            .position(position)
+            .locality(locality)
             .suggestion(message + wrongToken)
             .severity(WARNING)
             .build();
@@ -448,15 +474,13 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
     }
   }
 
-  @Nullable
-  private Position getPosition(Token childToken) {
-    return positionMapping.get(childToken);
+  private Optional<Locality> getLocality(Token childToken) {
+    return ofNullable(positionMapping.get(childToken));
   }
 
-  private void checkVariableDefinition(String name, @Nullable Position position) {
-    if (position == null) return;
+  private void checkVariableDefinition(String name, @Nonnull Locality locality) {
     if (!variables.contains(name)) {
-      reportVariableNotDefined(name, position, position); // starts and finishes in one token
+      reportVariableNotDefined(name, locality, locality); // starts and finishes in one token
     }
   }
 
@@ -472,12 +496,15 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
 
   private void checkVariableStructure(
       String parent, String child, Token childToken, Token parentToken) {
-    Position childPosition = getPosition(childToken);
-    Position parentPosition = getPosition(parentToken);
-    if (childPosition == null || parentPosition == null) return;
+    Optional<Locality> childLocality = getLocality(childToken);
+    Optional<Locality> parentLocality = getLocality(parentToken);
+
+    if (childLocality.isEmpty() || parentLocality.isEmpty()) return;
     if (!variables.parentContainsSpecificChild(parent, child)) {
       reportVariableNotDefined(
-          extractErrorStatementText(childToken, parentToken), childPosition, parentPosition);
+          extractErrorStatementText(childToken, parentToken),
+          childLocality.get(),
+          parentLocality.get());
     }
   }
 
@@ -488,27 +515,20 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
         .replaceAll(" +", " ");
   }
 
-  private void addUsage(SubContext<?> langContext, String name, @Nullable Position position) {
-    if (position == null) return;
-    langContext.addUsage(name.toUpperCase(), position);
+  private Locality getIntervalPosition(Locality start, Locality stop) {
+    return Locality.builder()
+        .uri(start.getUri())
+        .range(new Range(start.getRange().getStart(), stop.getRange().getEnd()))
+        .recognizer(CobolVisitor.class)
+        .build();
   }
 
-  private Position getIntervalPosition(Position start, Position stop) {
-    return new Position(
-        start.getDocumentURI(),
-        start.getStartPosition(),
-        stop.getStopPosition(),
-        start.getLine(),
-        start.getCharPositionInLine(),
-        start.getToken());
-  }
-
-  private void reportVariableNotDefined(String dataName, Position start, Position stop) {
+  private void reportVariableNotDefined(String dataName, Locality start, Locality stop) {
     SyntaxError error =
         SyntaxError.syntaxError()
             .suggestion(INVALID_DEF_MSG + dataName)
             .severity(INFO)
-            .position(getIntervalPosition(start, stop))
+            .locality(getIntervalPosition(start, stop))
             .build();
     LOG.debug("Syntax error by CobolVisitor#reportVariableNotDefined: " + error.toString());
     errors.add(error);
@@ -516,43 +536,46 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
 
   private void throwWarning(Token token) {
     MisspelledKeywordDistance.calculateDistance(token.getText().toUpperCase())
-        .ifPresent(correctWord -> reportMisspelledKeyword(correctWord, getPosition(token)));
+        .ifPresent(
+            correctWord ->
+                getLocality(token)
+                    .ifPresent(locality -> reportMisspelledKeyword(correctWord, locality)));
   }
 
-  private void reportMisspelledKeyword(String suggestion, Position position) {
-    if (position == null) return;
+  private void reportMisspelledKeyword(String suggestion, Locality locality) {
+    if (locality == null) return;
     SyntaxError error =
         SyntaxError.syntaxError()
             .suggestion(MISSPELLED_WORD + suggestion)
             .severity(WARNING)
-            .position(position)
+            .locality(locality)
             .build();
     LOG.debug("Syntax error by CobolVisitor#reportMisspelledKeyword: " + error.toString());
     errors.add(error);
   }
 
   private void areaAWarning(Token token) {
-    Position position = getPosition(token);
-    if (position == null) return;
-    if (position.getCharPositionInLine() > 10) {
-      throwException(token.getText(), position, AREA_A_WARNING_MSG);
-    }
+    getLocality(token)
+        .filter(it -> it.getRange().getStart().getCharacter() > 10)
+        .ifPresent(it -> throwException(token.getText(), it, AREA_A_WARNING_MSG));
   }
 
   private void areaBWarning(List<Token> tokenList) {
-    for (Token token : tokenList) {
-      Position position = getPosition(token);
-      if (position == null) continue;
-      int charPosition = position.getCharPositionInLine();
-      if (charPosition > 6 && charPosition < 11 && token.getChannel() != 1) {
-        throwException(token.getText(), position, AREA_B_WARNING_MSG);
-      }
-    }
+    tokenList.forEach(
+        token ->
+            getLocality(token)
+                .ifPresent(
+                    locality -> {
+                      int charPosition = locality.getRange().getStart().getCharacter();
+                      if (charPosition > 6 && charPosition < 11 && token.getChannel() != 1) {
+                        throwException(token.getText(), locality, AREA_B_WARNING_MSG);
+                      }
+                    }));
   }
 
   private void checkProgramName(Token token) {
     if (programName == null) {
-      throwException("", getPosition(token), "There is an issue with PROGRAM-ID paragraph");
+      getLocality(token).ifPresent(it -> throwException("", it, PROGRAM_ID_ISSUE_MSG));
     } else {
       checkProgramNameIdentical(token);
     }
@@ -561,7 +584,7 @@ public class CobolVisitor extends CobolParserBaseVisitor<Class> {
   private void checkProgramNameIdentical(Token token) {
     String text = PreprocessorStringUtils.trimQuotes(token.getText());
     if (!programName.equals(text)) {
-      throwException(programName, getPosition(token), IDENTICAL_PROGRAM_MSG);
+      getLocality(token).ifPresent(it -> throwException(programName, it, IDENTICAL_PROGRAM_MSG));
     }
   }
 }