+++ server/src/test/java/com/ca/lsp/cobol/usecases/TestCopybookCaching.java
@@ -18,20 +18,20 @@ package com.ca.lsp.cobol.usecases;
 
 import com.broadcom.lsp.cdi.EngineModule;
 import com.broadcom.lsp.cdi.module.databus.DatabusModule;
-import com.broadcom.lsp.domain.cobol.databus.api.CopybookRepository;
 import com.broadcom.lsp.domain.cobol.databus.api.DataBusBroker;
-import com.broadcom.lsp.domain.cobol.databus.model.CopybookStorable;
-import com.broadcom.lsp.domain.common.model.Position;
 import com.ca.lsp.cobol.positive.CobolText;
 import com.ca.lsp.cobol.service.mocks.MockCopybookService;
 import com.ca.lsp.cobol.service.mocks.MockCopybookServiceImpl;
+import com.ca.lsp.core.cobol.model.CopybookModel;
+import com.ca.lsp.core.cobol.model.Locality;
 import com.ca.lsp.core.cobol.preprocessor.sub.document.impl.CopybookResolutionProvider;
-import com.ca.lsp.core.cobol.preprocessor.sub.util.impl.MultiMapSerializableHelper;
 import com.google.common.collect.HashMultimap;
 import com.google.common.collect.Multimap;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import lombok.extern.slf4j.Slf4j;
+import org.eclipse.lsp4j.Position;
+import org.eclipse.lsp4j.Range;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -49,12 +49,15 @@ import static junit.framework.TestCase.assertTrue;
  */
 @Slf4j
 class TestCopybookCaching {
-  private final String COPYBOOK_NAME = "copy";
-  private final String COPYBOOK_CONTENT = "000230 77  REPORT-STATUS           PIC 99 VALUE ZERO.";
+  private static final String COPYBOOK_NAME = "copy";
+  private static final String COPYBOOK_CONTENT =
+      "000230 77  REPORT-STATUS           PIC 99 VALUE ZERO.";
 
-  private final Position POSITION_FIRST_OCCURRENCE = new Position(null, 0, 0, 0, 0, null);
-  private final Position POSITION_SECOND_OCCURRENCE = new Position(null, 10, 10, 10, 10, null);
-  private final Multimap<String, Position> paragraphDefinitions = HashMultimap.create();
+  private static final Locality FIRST_OCCURRENCE =
+      Locality.builder().range(new Range(new Position(0, 0), new Position(0, 0))).build();
+  private static final Locality SECOND_OCCURRENCE =
+      Locality.builder().range(new Range(new Position(10, 10), new Position(10, 10))).build();
+  private static final Multimap<String, Locality> paragraphDefinitions = HashMultimap.create();
 
   private DataBusBroker databus;
   private CopybookResolutionProvider resolution;
@@ -71,17 +74,12 @@ class TestCopybookCaching {
 
   private void initParagraphDefinitions() {
     String paragraphName = "PARNAME";
-    paragraphDefinitions.put(paragraphName, POSITION_FIRST_OCCURRENCE);
-    paragraphDefinitions.put(paragraphName, POSITION_SECOND_OCCURRENCE);
+    paragraphDefinitions.put(paragraphName, FIRST_OCCURRENCE);
+    paragraphDefinitions.put(paragraphName, SECOND_OCCURRENCE);
   }
 
   private void predefineCache() {
-    databus.storeData(
-        CopybookStorable.builder()
-            .name(COPYBOOK_NAME)
-            .uri(COPYBOOK_NAME)
-            .content(COPYBOOK_CONTENT)
-            .build());
+    databus.storeData(new CopybookModel(COPYBOOK_NAME, COPYBOOK_NAME, COPYBOOK_CONTENT));
   }
 
   private void initWorkspaceService() {
@@ -93,18 +91,6 @@ class TestCopybookCaching {
   @AfterEach
   void cleanup() {
     databus.invalidateCache();
-  }
-
-  @Test
-  void serializeMultiMapInString() {
-    LOG.info("Test serialization in String");
-    assertTrue(MultiMapSerializableHelper.serializeInString(paragraphDefinitions).length() > 0);
-  }
-
-  @Test
-  void serializeMultiMapInHashMap() {
-    LOG.info("Test serialization in HashMap");
-    assertTrue(MultiMapSerializableHelper.serializeInHashMap(paragraphDefinitions).size() > 0);
   }
 
   /** This test verifies that after the analysis a specific copybook retrieved from the cache */
@@ -132,14 +118,14 @@ class TestCopybookCaching {
     LOG.info(databus.printCache());
     runAnalysis("DID_OPEN");
     LOG.info(databus.printCache());
-    assertPositiveHitFromCache();
+    assertStoredInCache();
   }
 
   private void assertDidChangeAnalysisFromCache() {
     LOG.info(databus.printCache());
     runAnalysis("DID_CHANGE");
     LOG.info(databus.printCache());
-    assertPositiveHitFromCache();
+    assertStoredInCache();
   }
 
   private void assertDidOpenFromCopybookService() {
@@ -157,13 +143,7 @@ class TestCopybookCaching {
   }
 
   private void assertStoredInCache() {
-    assertTrue(databus.isStored(CopybookRepository.calculateUUID(COPYBOOK_NAME)));
-  }
-
-  private void assertPositiveHitFromCache() {
-    assertTrue(
-        databus.getData(CopybookRepository.calculateUUID(new StringBuilder(COPYBOOK_NAME))).getHit()
-            > 0);
+    assertTrue(databus.isStored(COPYBOOK_NAME));
   }
 
   private void runAnalysis(String syncType) {